version 4

+++ ComplexPlane.h
@@ -34,7 +34,7 @@ private:
 	Vector2f m_plane_size;
 	int m_zoomCount;
 	float m_aspetcRatio;
-	int countIterations(Vector2i corrd);
+	int countIterations(Vector2f corrd);
 	void iterationsToRGB(size_t count, Unit8& r, Unit8& g,Unit8& b);
 	Vector2f mapPixelToCoords(Vector2i mousePixel);
